Move `CPUTimer` into new `posix` mod
This technically unbreaks Windows, though we need to add a timer for it.
